Get back to privious solution without traits
