+++ src/stream/windows.rs
